requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
fredapi
python-dotenv
httpx
pydantic-settings
finance-datareader
